+++ application/views/timetable_all.php
@@ -1,0 +1,45 @@
+<h1>{pagetitle}</h1>
+
+<div class="row">
+  <div class="col-md-4">
+    <h2>Days</h2>
+    {daysofweek}
+    <dl class="dl-horizontal">
+      <dt>Day</dt><dd>{day}</dd>
+      <dt>Time</dt><dd>{time}</dd>
+      <dt>Course</dt><dd>{coursename}</dd>
+      <dt>Teacher</dt><dd>{teacher}</dd>
+      <dt>Location</dt><dd>{location}</dd>
+      <dt>Type</dt><dd>{classtype}</dd>
+    </dl>
+    {/daysofweek}
+  </div>
+
+  <div class="col-md-4">
+    <h2>Courses</h2>
+    {courses}
+    <dl class="dl-horizontal">
+      <dt>Day</dt><dd>{day}</dd>
+      <dt>Time</dt><dd>{time}</dd>
+      <dt>Course</dt><dd>{coursename}</dd>
+      <dt>Teacher</dt><dd>{teacher}</dd>
+      <dt>Location</dt><dd>{location}</dd>
+      <dt>Type</dt><dd>{classtype}</dd>
+    </dl>
+    {/courses}
+  </div>
+
+  <div class="col-md-4">
+    <h2>Periods</h2>
+    {periods}
+    <dl class="dl-horizontal">
+      <dt>Day</dt><dd>{day}</dd>
+      <dt>Time</dt><dd>{time}</dd>
+      <dt>Course</dt><dd>{coursename}</dd>
+      <dt>Teacher</dt><dd>{teacher}</dd>
+      <dt>Location</dt><dd>{location}</dd>
+      <dt>Type</dt><dd>{classtype}</dd>
+    </dl>
+    {/periods}
+  </div>
+</div>