+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 All notable changes to this project will be documented to this file.
 Change Log history will be recorded in descending order
 
+## 0.1
+- Update Timetable Model to find duplicates
+- Update search view to display an error if duplicates are found
+
+## 0.0.9
+- Add a search form to header
+- Add search page to show results
+- Load all timetable entries and display them on index page
+
+## 0.0.8-alpha
+Added a basic template and layout files for the web view
+Tried loading the model but it broke
+
+## 0.0.7
+** Added Files **
+- Timetable.php
+
+## 0.0.6
+** Added Files **
+- timetable.dtd
+- timetable.xml
+
+** Deleted Files **
+- course.xml
+- period.xml
+- schedule.xml
+- days.xml
+
 ## 0.0.5
 Update case of the data folder to match name convention